# segment_appender: truncate in addition to allocate