@vbotbuildovich 2024-05-23
<!--NEW FAILURES-->
new failures in https://buildkite.com/redpanda/redpanda/builds/49484#018fa51d-8576-4866-bcb5-bde575599b23:
```
"rptest.tests.archival_test.ArchivalTest.test_one_node_reconnect.cloud_storage_type=CloudStorageType.ABS"
"rptest.tests.archival_test.ArchivalTest.test_write.cloud_storage_type=CloudStorageType.ABS"
"rptest.tests.archival_test.ArchivalTest.test_connection_flicker.cloud_storage_type=CloudStorageType.ABS"
```
<!--07c8cbd0-bb5a-4494-ab17-8554d213c0b2-->