@vbotbuildovich 2024-05-23
<!--RETRIED FAILURES-->
ducktape was retried in https://buildkite.com/redpanda/redpanda/builds/49484#018fa51d-8578-41c6-989e-2844b6df16d8
<!--cc74bbd3-56f7-4114-97b2-e11e033755f5-->