@vbotbuildovich 2024-05-23
<!--NEW FAILURES-->
new failures in https://buildkite.com/redpanda/redpanda/builds/49489#018fa5e7-998f-4353-8abd-2913b06ceaff:
```
"rptest.tests.topic_recovery_test.TopicRecoveryTest.test_many_partitions.cloud_storage_type=CloudStorageType.ABS.check_mode=check_manifest_and_segment_metadata"
"rptest.tests.topic_recovery_test.TopicRecoveryTest.test_many_partitions.cloud_storage_type=CloudStorageType.S3.check_mode=check_manifest_existence"
```
<!--54483447-9834-4c0c-9606-f93cae110519-->