@vbotbuildovich 2024-05-23
<!--RETRIED FAILURES-->
ducktape was retried in https://buildkite.com/redpanda/redpanda/builds/49489#018fa5d0-3e9a-4032-8be7-2fa26413773f
<!--fc78d854-1852-46fb-8c34-3a78be459708-->