@vbotbuildovich 2024-05-28
<!--RETRIED FAILURES-->
ducktape was retried in https://buildkite.com/redpanda/redpanda/builds/49598#018fbf16-947c-4a6f-a54e-550f8fcc8ad7
<!--2907a18b-1bfc-4ad2-a903-9467e780442d-->