@StephanDollberg 2024-05-28
> LGTM. Did you notice increase in test duration? This new compute storage doesn't seem very lightweight and wonder if it would make more sense to adjust tests to fallocate less/not rely on that precise segment sizing information.

Doesn't seem very noticable. Maybe very little.

> more sense to adjust tests to fallocate less/not rely on that precise segment sizing information

I am very open to ideas but can't immediately see how we could improve it. The size checking already feels like a simplifying hack (all bits could be garbled). I wouldn't want to change all the tests to use smaller segment sizes etc as I prefer things to run in default configs where possible.