@StephanDollberg 2024-05-30
I have pushed some changes now that speed up the size computation in `data_checksum`. Haven't tested all tests yet so will need to see whether everything still passes.