@StephanDollberg 2024-06-08
> Could I request having a DNM in PR then

Ah yeah sure good point, will add next time.

> Should we revert?

Nah not worth the hassle. I think we should have a few data points at least now since we made the throughput changes.